test(relay): extract shared control-plane session mock
Replaces the mock Session/Subscriber pairs duplicated across the
unsubscribe, malformed-track, and fetch-ingest tests with a single
recording mock in core::mocks, per the shared-Arrange testing rule.
