+++ relay/src/modules/sequences/unsubscribe.rs
@@ -108,24 +108,16 @@ impl Unsubscribe {
 
 #[cfg(test)]
 mod tests {
-    use std::sync::{Arc, Mutex};
-
     use tokio::sync::mpsc;
 
     use super::*;
     use crate::modules::{
-        core::{
-            data_receiver::fetch_receiver::UpstreamFetchReceiver,
-            data_receiver::receiver::DataReceiver, handler::publish::SubscribeOption,
-            publisher::Publisher, session::Session, session_event::MoqtSessionEvent,
-            subscriber::Subscriber, subscription::UpstreamSubscription,
-        },
+        core::mocks::{RecordedControlMessages, session_repository_with_mock_control_session},
         enums::ContentExists,
         sequences::tables::{
             hashmap_table::InMemoryLocalPubSubDirectory,
             table::{ActiveUpstreamSubscription, UpstreamSubscriptionKey},
         },
-        session_repository::SessionRepository,
         types::TrackKey,
     };
 
@@ -142,84 +134,6 @@ mod tests {
         }
     }
 
-    struct MockSubscriber {
-        unsubscribed_request_ids: Arc<Mutex<Vec<u64>>>,
-    }
-
-    #[async_trait::async_trait]
-    impl Subscriber for MockSubscriber {
-        async fn send_subscribe(
-            &mut self,
-            _track_namespace: String,
-            _track_name: String,
-            _option: SubscribeOption,
-        ) -> anyhow::Result<UpstreamSubscription> {
-            unimplemented!("not used in unsubscribe tests")
-        }
-
-        async fn send_unsubscribe(&self, subscribe_id: u64) -> anyhow::Result<()> {
-            self.unsubscribed_request_ids
-                .lock()
-                .unwrap()
-                .push(subscribe_id);
-            Ok(())
-        }
-
-        async fn send_unsubscribe_namespace(&self, _namespace: String) -> anyhow::Result<()> {
-            Ok(())
-        }
-
-        async fn create_data_receiver(
-            &mut self,
-            _subscription: &UpstreamSubscription,
-        ) -> anyhow::Result<DataReceiver> {
-            unimplemented!("not used in unsubscribe tests")
-        }
-
-        async fn send_fetch(
-            &mut self,
-            _track_namespace: String,
-            _track_name: String,
-            _start_location: moqt::Location,
-            _end_location: moqt::Location,
-            _option: moqt::FetchOption,
-        ) -> anyhow::Result<moqt::FetchHandle> {
-            unimplemented!("not used in unsubscribe tests")
-        }
-
-        async fn create_fetch_receiver(
-            &mut self,
-            _handle: &moqt::FetchHandle,
-        ) -> anyhow::Result<Box<dyn UpstreamFetchReceiver>> {
-            unimplemented!("not used in unsubscribe tests")
-        }
-
-        async fn send_fetch_cancel(&self, _request_id: u64) -> anyhow::Result<()> {
-            unimplemented!("not used in unsubscribe tests")
-        }
-    }
-
-    struct MockSession {
-        unsubscribed_request_ids: Arc<Mutex<Vec<u64>>>,
-    }
-
-    #[async_trait::async_trait]
-    impl Session for MockSession {
-        fn as_publisher(&self) -> Box<dyn Publisher> {
-            unimplemented!("not used in unsubscribe tests")
-        }
-
-        fn as_subscriber(&self) -> Box<dyn Subscriber> {
-            Box::new(MockSubscriber {
-                unsubscribed_request_ids: self.unsubscribed_request_ids.clone(),
-            })
-        }
-
-        async fn receive_moqt_session_event(&self) -> anyhow::Result<MoqtSessionEvent> {
-            std::future::pending().await
-        }
-    }
-
     struct TestContext {
         table: InMemoryLocalPubSubDirectory,
         forwarder: ControlMessageForwarder,
@@ -227,7 +141,7 @@ mod tests {
         ingress_receiver: mpsc::Receiver<IngressCommand>,
         egress_sender: mpsc::Sender<EgressCommand>,
         egress_receiver: mpsc::Receiver<EgressCommand>,
-        unsubscribed_request_ids: Arc<Mutex<Vec<u64>>>,
+        recorded: RecordedControlMessages,
     }
 
     async fn setup(
@@ -260,22 +174,9 @@ mod tests {
             ));
         }
 
-        let unsubscribed_request_ids = Arc::new(Mutex::new(Vec::new()));
-        let mut repository = SessionRepository::new();
-        let (session_event_sender, _session_event_receiver) = mpsc::unbounded_channel();
-        repository
-            .add_client(
-                PUBLISHER_SESSION,
-                Box::new(MockSession {
-                    unsubscribed_request_ids: unsubscribed_request_ids.clone(),
-                }),
-                session_event_sender,
-                tracing::Span::none(),
-            )
-            .await;
-        let forwarder = ControlMessageForwarder {
-            repository: Arc::new(tokio::sync::Mutex::new(repository)),
-        };
+        let (repository, recorded) =
+            session_repository_with_mock_control_session(PUBLISHER_SESSION).await;
+        let forwarder = ControlMessageForwarder { repository };
 
         let (ingress_sender, ingress_receiver) = mpsc::channel(8);
         let (egress_sender, egress_receiver) = mpsc::channel(8);
@@ -286,7 +187,7 @@ mod tests {
             ingress_receiver,
             egress_sender,
             egress_receiver,
-            unsubscribed_request_ids,
+            recorded,
         }
     }
 
@@ -306,10 +207,13 @@ mod tests {
 
     #[tokio::test]
     async fn last_subscriber_forwards_upstream_unsubscribe_and_stops_ingress() {
+        // Arrange
         let mut ctx = setup(UpstreamSubscriptionOrigin::Subscribe, &[(100, 10)]).await;
 
+        // Act
         run_unsubscribe(&ctx, 100, 10).await;
 
+        // Assert: the reader stops, the upstream is unsubscribed, ingress stops.
         match ctx.egress_receiver.try_recv() {
             Ok(EgressCommand::StopReader {
                 subscriber_session_id,
@@ -321,7 +225,7 @@ mod tests {
             other => panic!("Expected StopReader, got {:?}", other.is_ok()),
         }
         assert_eq!(
-            *ctx.unsubscribed_request_ids.lock().unwrap(),
+            *ctx.recorded.unsubscribed_request_ids.lock().unwrap(),
             vec![UPSTREAM_REQUEST_ID]
         );
         match ctx.ingress_receiver.try_recv() {
@@ -338,44 +242,71 @@ mod tests {
 
     #[tokio::test]
     async fn remaining_subscribers_keep_upstream_subscription() {
+        // Arrange
         let mut ctx = setup(
             UpstreamSubscriptionOrigin::Subscribe,
             &[(100, 10), (101, 11)],
         )
         .await;
 
+        // Act
         run_unsubscribe(&ctx, 100, 10).await;
 
+        // Assert: only the reader stops; the upstream subscription survives.
         assert!(matches!(
             ctx.egress_receiver.try_recv(),
             Ok(EgressCommand::StopReader { .. })
         ));
-        assert!(ctx.unsubscribed_request_ids.lock().unwrap().is_empty());
+        assert!(
+            ctx.recorded
+                .unsubscribed_request_ids
+                .lock()
+                .unwrap()
+                .is_empty()
+        );
         assert!(ctx.ingress_receiver.try_recv().is_err());
     }
 
     #[tokio::test]
     async fn publish_origin_keeps_upstream_subscription() {
+        // Arrange
         let mut ctx = setup(UpstreamSubscriptionOrigin::Publish, &[(100, 10)]).await;
 
+        // Act
         run_unsubscribe(&ctx, 100, 10).await;
 
+        // Assert: only the reader stops; the publisher pushed the track.
         assert!(matches!(
             ctx.egress_receiver.try_recv(),
             Ok(EgressCommand::StopReader { .. })
         ));
-        assert!(ctx.unsubscribed_request_ids.lock().unwrap().is_empty());
+        assert!(
+            ctx.recorded
+                .unsubscribed_request_ids
+                .lock()
+                .unwrap()
+                .is_empty()
+        );
         assert!(ctx.ingress_receiver.try_recv().is_err());
     }
 
     #[tokio::test]
     async fn unknown_subscription_sends_no_commands() {
+        // Arrange
         let mut ctx = setup(UpstreamSubscriptionOrigin::Subscribe, &[]).await;
 
+        // Act
         run_unsubscribe(&ctx, 100, 10).await;
 
+        // Assert
         assert!(ctx.egress_receiver.try_recv().is_err());
-        assert!(ctx.unsubscribed_request_ids.lock().unwrap().is_empty());
+        assert!(
+            ctx.recorded
+                .unsubscribed_request_ids
+                .lock()
+                .unwrap()
+                .is_empty()
+        );
         assert!(ctx.ingress_receiver.try_recv().is_err());
     }
 }
